requirements.txt
flask
cryptography
mysql-connector-python
